Cleanup metadata generation
